backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
transformers>=4.39.0
torch>=2.1.0
google-generativeai>=0.8.5
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0